backend/ecommerce/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email]==2.7.1
pydantic-settings==2.3.0
redis==5.0.4
celery==5.4.0
httpx==0.27.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
prophet==1.1.5
kaggle==1.6.14
python-dotenv==1.0.1
stripe==9.12.0
pytest==8.2.2
pytest-asyncio==0.23.7
